requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
future==0.18.2
geetest==3.2.1
geetest-django==0.0.1
haystack==0.42
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
pefile==2023.2.7
PyJWT==1.7.1
PyMySQL==1.0.2
python-ptrace==0.9.8
pytz==2022.6
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
